refactor: move token generator
