+++ internal/app/domain/services/token_generator.go
@@ -1,4 +1,4 @@
-package auth
+package services
 
 import "errors"
