+++ internal/app/infrastructure/services/stub/token_generator.go
